simplify root layout by using components for header and sidebar
